Hard code Fortran GRIB types so grib_types doesn't need -lc
On Frontier this program gives a segfault when built with
grib_fortran_kinds.c, because doing so requires linking with -lc, which
gives a program that always segfaults for unknown reasons. In the mean
time it's easier just to hard code the output of this program so it
doesn't need the C function at all.